Agregar /api en las rutas y separacion de personas y usuarios
