+++ src/index.js
@@ -5,6 +5,7 @@ import { dirname, join } from "path";
 import { fileURLToPath } from "url";
 import { dbOptions } from "./database.js";
 import UserRoutes from "./routes/users.js";
+import PersonsRoutes from './routes/persons.js'
 //para conectarlo a la base de datos
 import mysql from "mysql2";
 import myconn from "express-myconnection";
@@ -25,7 +26,8 @@ app.use(express.json());
 app.use(myconn(mysql, dbOptions, "single")); //conexion base de datos
 
 // Rutas
-app.use(UserRoutes);
+app.use("/api", UserRoutes);
+app.use("/api", PersonsRoutes);
 // app.use('RUTA')
 
 // Archivos estáticos
